(update: backend) user, cats and favorites relations

+++ backend/src/cats/cats.service.ts
@@ -1,4 +1,9 @@
-import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
+import {
+  HttpException,
+  HttpStatus,
+  Injectable,
+  UnauthorizedException,
+} from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { Cat } from "./entities/cat.entity";
@@ -10,20 +15,25 @@ export class CatsService {
     private catsRepository: Repository<Cat>
   ) {}
 
-  // Create a new cat
-  async create(createCatDto: CreateCatDto): Promise<Cat> {
-    const newCat = this.catsRepository.create(createCatDto);
+  async create(createCatDto: CreateCatDto, userId: number): Promise<Cat> {
+    const newCat = this.catsRepository.create({
+      ...createCatDto,
+      user: { id: userId },
+    });
     return this.catsRepository.save(newCat);
   }
 
   // Find all cats
   async findAll(): Promise<Cat[]> {
-    return this.catsRepository.find();
+    return this.catsRepository.find({ relations: ["user"] });
   }
 
   // Find a cat by ID
   async findOne(id: number, action: "update" | "delete" | "get"): Promise<Cat> {
-    const cat = await this.catsRepository.findOne({ where: { id } });
+    const cat = await this.catsRepository.findOne({
+      where: { id },
+      relations: ["user"],
+    });
     if (!cat) {
       throw new HttpException(
         `Cannot ${action} cat. Cat with ID ${id} not found.`,
@@ -43,8 +53,18 @@ export class CatsService {
   }
 
   // Remove a cat
-  async remove(id: number): Promise<{ message: string }> {
+  async remove(id: number, userId: number): Promise<{ message: string }> {
+    // Find the cat to remove
     const catToRemove = await this.findOne(id, "delete");
+
+    // Check if the user is authorized to delete the cat
+    if (catToRemove.user.id !== userId) {
+      throw new UnauthorizedException(
+        "You are not authorized to delete this cat"
+      );
+    }
+
+    // Remove the cat
     await this.catsRepository.remove(catToRemove);
     return { message: `Cat deleted successfully` };
   }
